DECODE Auto & Tele V1
Added the first prototypes for the autonomous and teleop programs for the DECODE season.

Much of it is still untested, but they should work adequately with just a few tweaks. I will document all of the changes from last year once these programs are finalized for the first competition.

I am only committing these files right now in order to store a backup as I create a new project in Android Studio and test what is currently wrong with the Lynx modules (I'm not sure if it's a software or hardware problem)